Add ability to treat some DELTA metrics as counter rather than gauge
In order for us to be able to do some performance math on these in
prometheus, we need to have counters.

This code adds a CLI flag to define a substring (underscore delimited)
that indicates a DELTA metric should be treated as a counter.

This is accomplished by caching the "current" running value in this case
and exporting that value as a counter.

The helper code for this is in the collector namespace (I am using the
fnv code and followed a similar pattern in adding it here).

Currently this code only handles INT64 DELTAs.  I've identified a
handful of DOUBLE DELTAs that could require similar treatment.  That's a
TODO.