[SRE-5550] Remove no longer useful log line
Because this else now encapsulates several possibilities (and there was originally no logging here)
I think this should go away as it's potentially confusing.

The possibilities:

- INT64 or DOUBLE is not DELTA
- flag to treat all DELTAs as counters is not set
- both

Because the log is saying "hey this is not a DELTA counter" it's not
really capturing the essence of what could take that path.